test(condition): Implement condition test to know that a test can be given a condition
This test explains that a test can be given a condition where if the
condition is met, the test will be run and if not, the test will be
stopped. For example, if we have users with different types of levels so
that each level will get different resources. In this test we can use
manual conditioning with if-else and then throw a TestAbortedException,
and can use the Assumptions class static functions and also use an
annotations.